requirements.txt
torch
transformers
python-dotenv
discord.py
accelerate
